pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-primitives-gateway"
version = "0.1.0"
description = "Agentic primitives gateway - unified API for agentic primitives"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
mem0 = [
    "mem0ai>=0.1.0",
    "pymilvus>=2.4.0",
]
agentcore = [
    "bedrock-agentcore>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-sdk-extension-aws>=2.0.0",
    "aws-opentelemetry-distro>=0.10.0",
    "playwright>=1.40.0",
]
langfuse = [
    "langfuse>=3.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "types-PyYAML>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_primitives_gateway"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "client/src", "tests", "client/tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line-too-long (formatter handles code; decorative comments are fine)
]

[tool.ruff.lint.isort]
known-first-party = ["agentic_primitives_gateway", "agentic_primitives_gateway_client"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = false
mypy_path = "src:client/src"
packages = ["agentic_primitives_gateway", "agentic_primitives_gateway_client"]

[[tool.mypy.overrides]]
module = [
    "mem0.*",
    "bedrock_agentcore.*",
    "opentelemetry.*",
    "aws_opentelemetry.*",
    "playwright.*",
    "langfuse.*",
    "pymilvus.*",
    "ddtrace.*",
    "boto3.*",
    "botocore.*",
    "amazon.*",
]
ignore_missing_imports = true
